[refactor] 新增 AvatarBlock component
